requirements.txt
open-interpreter
python-dotenv
pyperclip
pyautogui
pywinctl
opencv-python
pytesseract 
pillow 
pandas